// data_frontend/app/form/types.ts
export interface FormFormData {
  projectName: string;
  formTheme: string;
  dateOfIdentification: string;
  location: string;
  gembaUnit: string;
  category: string;
  subCategory: string;
  department: string;
  currentSituation: string;
  rootCause: string;
  actionTaken: string;
  standardization: string;
  dateOfCompletion: string;
  beforePictures: File[];
  afterPictures: File[];
}